Optimized struct data casting
The casting iteration now is now performed on the prop list, rather than
on the input data. This removes the need for `getDefaults()` which
previously iterated over the properties and maps them to their default
values. This change effectively removes that redundant loop over the
properties, making struct construction much more performant.