Some refactoring ...
Struct prop types are no longer cached in a class static var.
This caused extended classes to share the cache, which is bad.

Tests now create anonymous classes to better simulate expected consumer
usage, which is to extend the `Struct` class. As such, the stub was no
longer required.

A mock prop type class was created to facilitate test authoring (since
we can't inject a mock builder into the class - mocked prop types need
to be available before the ctor is called, and anon classes
automatically call the ctor).

Most of the tests have also been revised to be more sensible to real
world expectations.

Finally, the equivalence method is now called `equals()` and only takes
two arguments to reduce complexity. Consumers can easily use this method
with `array_reduce()` to check for equivalence of multiple instances.